Banish source related fns to the lower depths of module hell
