+++ proxy/src/coco.rs
@@ -18,8 +18,11 @@ use radicle_keystore::{Keypair, Keystore, SecretKeyExt};
 
 use crate::error;
 
-mod types;
-pub use types::*; // TODO: make explicit
+mod source;
+pub use source::{
+    blob, branches, commit, commits, local_branches, tags, tree, Blob, BlobContent, Branch, Commit,
+    Info, ObjectType, Person, Tag, Tree, TreeEntry,
+};
 
 /// The set of capabilities necessary for interacting with `radicle-link`.
 #[async_trait]
@@ -380,262 +383,12 @@ where
             let path = format!("{}/{}/{}", root, "repos", info.0);
             std::fs::create_dir_all(path.clone())?;
 
-            init_repo(path.clone())?;
+            source::init_repo(path.clone())?;
             self.init_project(&owner, &path, info.0, info.1, info.2)?;
         }
 
         Ok(())
     }
-}
-
-/// Returns the [`Blob`] for a file at `revision` under `path`.
-///
-/// # Errors
-///
-/// Will return [`error::Error`] if the project doesn't exist or a surf interaction fails.
-pub fn blob<'a>(
-    browser: &mut surf::git::Browser<'a>,
-    revision: String,
-    maybe_path: Option<String>,
-) -> Result<Blob, error::Error> {
-    let path = maybe_path.clone().unwrap_or_default();
-
-    // Best effort to guess the revision.
-    browser.revspec(&revision)?;
-
-    let root = browser.get_directory()?;
-
-    let p = surf::file_system::Path::from_str(&path)?;
-
-    let file = root
-        .find_file(p.clone())
-        .ok_or_else(|| error::Error::PathNotFound)?;
-
-    let mut commit_path = surf::file_system::Path::root();
-    commit_path.append(p.clone());
-
-    let last_commit = browser.last_commit(commit_path)?.map(|c| Commit::from(&c));
-    let (_rest, last) = p.split_last();
-    let content = match std::str::from_utf8(&file.contents) {
-        Ok(content) => BlobContent::Ascii(content.to_string()),
-        Err(_) => BlobContent::Binary,
-    };
-
-    Ok(Blob {
-        content,
-        info: Info {
-            name: last.to_string(),
-            object_type: ObjectType::Blob,
-            last_commit,
-        },
-        path: maybe_path.unwrap_or(last.to_string()),
-    })
-}
-
-/// Given a project id to a repo returns the list of branches.
-///
-/// # Errors
-///
-/// Will return [`error::Error`] if the project doesn't exist or the surf interaction fails.
-pub fn branches<'a>(browser: &surf::git::Browser<'a>) -> Result<Vec<Branch>, error::Error> {
-    let mut branches = browser
-        .list_branches(None)?
-        .into_iter()
-        .map(|b| Branch(b.name.name().to_string()))
-        .collect::<Vec<Branch>>();
-
-    branches.sort();
-
-    Ok(branches)
-}
-
-/// Given a path to a repo returns the list of branches.
-///
-/// # Errors
-///
-/// Will return [`error::Error`] if the repository doesn't exist.
-pub fn local_branches(repo_path: &str) -> Result<Vec<Branch>, error::Error> {
-    let repo = surf::git::Repository::new(repo_path)?;
-    let browser = surf::git::Browser::new(&repo)?;
-    let mut branches = browser
-        .list_branches(None)?
-        .into_iter()
-        .map(|b| Branch(b.name.name().to_string()))
-        .collect::<Vec<Branch>>();
-
-    branches.sort();
-
-    Ok(branches)
-}
-
-/// Retrieves the [`Commit`] for the given `sha1`.
-///
-/// # Errors
-///
-/// Will return [`error::Error`] if the project doesn't exist or the surf interaction fails.
-pub fn commit<'a>(
-    browser: &mut surf::git::Browser<'a>,
-    sha1: &str,
-) -> Result<Commit, error::Error> {
-    browser.commit(surf::git::Oid::from_str(sha1)?)?;
-
-    let history = browser.get();
-    let commit = history.first();
-
-    Ok(Commit::from(commit))
-}
-
-/// Retrieves the [`Commit`] history for the given `branch`.
-///
-/// # Errors
-///
-/// Will return [`error::Error`] if the project doesn't exist or the surf interaction fails.
-pub fn commits<'a>(
-    browser: &mut surf::git::Browser<'a>,
-    branch: &str,
-) -> Result<Vec<Commit>, error::Error> {
-    browser.branch(surf::git::BranchName::new(branch))?;
-
-    let commits = browser.get().iter().map(Commit::from).collect();
-
-    Ok(commits)
-}
-
-/// Retrieves the list of [`Tag`] for the given project `id`.
-///
-/// # Errors
-///
-/// Will return [`error::Error`] if the project doesn't exist or the surf interaction fails.
-pub fn tags<'a>(browser: &surf::git::Browser<'a>) -> Result<Vec<Tag>, error::Error> {
-    let tag_names = browser.list_tags()?;
-
-    let mut tags: Vec<Tag> = tag_names
-        .into_iter()
-        .map(|tag_name| Tag(tag_name.name().to_string()))
-        .collect();
-
-    tags.sort();
-
-    Ok(tags)
-}
-
-/// Retrieve the [`Tree`] for the given `revision` and directory `prefix`.
-///
-/// # Errors
-///
-/// Will return [`error::Error`] if any of the surf interactions fail.
-/// TODO(fintohaps): default branch fall back from Browser
-pub fn tree<'a>(
-    browser: &mut surf::git::Browser<'a>,
-    revision: String,
-    maybe_prefix: Option<String>,
-) -> Result<Tree, error::Error> {
-    let prefix = maybe_prefix.unwrap_or_default();
-
-    browser.revspec(&revision)?;
-
-    let path = if prefix == "/" || prefix == "" {
-        surf::file_system::Path::root()
-    } else {
-        surf::file_system::Path::from_str(&prefix)?
-    };
-
-    let root_dir = browser.get_directory()?;
-    let prefix_dir = if path.is_root() {
-        root_dir
-    } else {
-        root_dir
-            .find_directory(path.clone())
-            .ok_or_else(|| error::Error::PathNotFound)?
-    };
-    let mut prefix_contents = prefix_dir.list_directory();
-    prefix_contents.sort();
-
-    let entries_results: Result<Vec<TreeEntry>, error::Error> = prefix_contents
-        .iter()
-        .map(|(label, system_type)| {
-            let entry_path = if path.is_root() {
-                surf::file_system::Path::new(label.clone())
-            } else {
-                let mut p = path.clone();
-                p.push(label.clone());
-                p
-            };
-            let mut commit_path = surf::file_system::Path::root();
-            commit_path.append(entry_path.clone());
-
-            let info = Info {
-                name: label.to_string(),
-                object_type: match system_type {
-                    surf::file_system::SystemType::Directory => ObjectType::Tree,
-                    surf::file_system::SystemType::File => ObjectType::Blob,
-                },
-                last_commit: None,
-            };
-
-            Ok(TreeEntry {
-                info,
-                path: entry_path.to_string(),
-            })
-        })
-        .collect();
-
-    let mut entries = entries_results?;
-
-    // We want to ensure that in the response Tree entries come first. `Ord` being derived on
-    // the enum ensures Variant declaration order.
-    //
-    // https://doc.rust-lang.org/std/cmp/trait.Ord.html#derivable
-    entries.sort_by(|a, b| a.info.object_type.cmp(&b.info.object_type));
-
-    let last_commit = if path.is_root() {
-        Some(Commit::from(browser.get().first()))
-    } else {
-        None
-    };
-    let name = if path.is_root() {
-        "".into()
-    } else {
-        let (_first, last) = path.split_last();
-        last.to_string()
-    };
-    let info = Info {
-        name,
-        object_type: ObjectType::Tree,
-        last_commit,
-    };
-
-    Ok(Tree {
-        path: prefix,
-        entries,
-        info,
-    })
-}
-
-/// Initialize a [`radicle_surf::git::git2::Repository`] at the given path.
-///
-/// # Errors
-///
-/// Will return [`error::Error`] if any of the git interactions break.
-pub fn init_repo(path: String) -> Result<(), error::Error> {
-    let repo = git2::Repository::init(path)?;
-
-    // First use the config to initialize a commit signature for the user.
-    let sig = repo.signature()?;
-    // Now let's create an empty tree for this commit
-    let tree_id = {
-        let mut index = repo.index()?;
-
-        // For our purposes, we'll leave the index empty for now.
-        index.write_tree()?
-    };
-    let tree = repo.find_tree(tree_id)?;
-    // Normally creating a commit would involve looking up the current HEAD
-    // commit and making that be the parent of the initial commit, but here this
-    // is the first commit so there will be no parent.
-    repo.commit(Some("HEAD"), &sig, &sig, "Initial commit", &tree, &[])?;
-
-    Ok(())
 }
 
 /// Intermediate type until we have a proper implementation.
